Refactor out common components
