+++ src/components/Event/index.ts
@@ -1,7 +1,7 @@
 import { Stream } from 'xstream';
 import { div, article, h2, h3, button, p, VNode, DOMSource } from '@cycle/dom';
 import isolate from '@cycle/isolate';
-import { DevdayEvent} from '../../../../definitions';
+import { DevdayEvent} from '../../definitions';
 import sampleCombine from 'xstream/extra/sampleCombine';
 import * as moment from 'moment';
 import * as marked from 'marked';
