+++ src/components/Event/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/devday.scss';
+@import '../../styles/devday.scss';
 
 .event {
   @extend .mdl-cell;
